Notebook testing script.
